backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
twilio
google-generativeai
python-multipart
gtts
